updating count

+++ Bubblesort.java
@@ -10,7 +10,7 @@ import java.util.Random;
 import java.util.Scanner;
 public class Bubblesort
 {
-    public static int count = 0;
+    public static long count = 0;
     public static void main(String[] args)
     {
         Scanner sc = new Scanner(System.in);
