+++ Mergesort.java
@@ -11,7 +11,7 @@ import java.util.Random;
 import java.lang.Math;
 public class Mergesort
 {
-    public static int count = 0;
+    public static long count = 0;
     public static void main(String[] args)
     {
         Scanner sc = new Scanner(System.in);
